+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - PostgreSQL: Display description of system variables
 - PostgreSQL: Avoid warning about crdb_version (bug #924, regression from 5.0.5)
 - CSS: Sticky table headers (bug #918)
+- CSS: Allow more custom styles with dark mode (bug #925)
 - IMAP: New plugin driver created for fun
 
 ## Adminer 5.0.6 (released 2025-03-17)